Separate projects and experiments
