+++ api/experiments/serialiazers.py
@@ -1,0 +1,23 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, division, print_function
+
+from rest_framework import serializers
+
+from experiments.models import Experiment
+
+
+class ExperimentSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Experiment
+        fields = ('id', 'name', 'created_at', 'updated_at', )
+
+
+class ExperimentDetailSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Experiment
+        fields = '__all__'
+
+
+class StatusSerializer(serializers.Serializer):
+    job_id = serializers.CharField()
+    status = serializers.CharField()
