+++ api/libs/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ProjectsConfig(AppConfig):
+    name = 'Libs'
+    verbose_name = 'Libs'
