+++ api/projects/apps.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, division, print_function
+
+from django.apps import AppConfig
+
+
+class ProjectsConfig(AppConfig):
+    name = 'Projects'
+    verbose_name = 'Projects'
