+++ api/projects/serialiazers.py
@@ -3,13 +3,8 @@ from __future__ import absolute_import, division, print_function
 
 from rest_framework import serializers
 
-from projects.models import Project, Experiment
-
-
-class ExperimentSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Experiment
-        fields = '__all__'
+from experiments.serialiazers import ExperimentSerializer
+from projects.models import Project
 
 
 class ProjectSerializer(serializers.ModelSerializer):
